[tag_edit] separate responsibility for tags, source, owner, and lock
