+++ ext/post_lock/test.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shimmie2;
+
+class PostLockTest extends ShimmiePHPUnitTestCase
+{
+}
