+++ ext/post_source/test.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shimmie2;
+
+class PostSourceTest extends ShimmiePHPUnitTestCase
+{
+    public function testSourceEdit(): void
+    {
+        $this->log_in_as_user();
+        $image_id = $this->post_image("tests/pbx_screenshot.jpg", "pbx");
+        $image = Image::by_id($image_id);
+
+        send_event(new SourceSetEvent($image, "example.com"));
+        send_event(new SourceSetEvent($image, "http://example.com"));
+
+        $this->get_page("post/view/$image_id");
+        $this->assert_text("example.com");
+    }
+}
